+++ NEWS.md
@@ -1,5 +1,7 @@
 # daedalus.api 0.1.6
 
+This patch version updates cost and results outputs (PR #47).
+
 - Updates to function `get_nested_costs()` and JSON schema `scenarioCosts.json` to return costs as an array of metric-value pairs; this allows costs to be expressed in different units.
 
 - Updates to results schema `scenarioResults.json` to return the age-specific VSL for countries.
@@ -10,7 +12,7 @@
 
 - Moved `cost_item()` function from inline helper to internal package function.
 
-- Updates to tests helper functions.
+- Updates to tests helper functions; added helper tests for nested list outputs in `R/tests.R`.
 
 # daedalus.api 0.1.5